bfdr(feat): add new messaging class to support birth and fetal death messaging
