+++ projects/BFDR.Tests/FetalDeathRecord_Should.cs
@@ -28,7 +28,7 @@ namespace BFDR.Tests
         SetterFetalDeathRecord.StateLocalIdentifier1 = "0000033";
         Assert.Equal("0000033", SetterFetalDeathRecord.StateLocalIdentifier1);
         // TODO: Delivery year is not implemented yet
-        // SetterFetalDeathRecord.DeliveryYear = 2020;
+        SetterFetalDeathRecord.DeliveryYear = 2020;
         // SetterFetalDeathRecord.CertificateNumber = "767676";
         // Assert.Equal("2020XX767676", SetterFetalDeathRecord.RecordIdentifier);
         SetterFetalDeathRecord.BirthLocationJurisdiction = "WY";
